grabber/requirements.txt
fastapi==0.118.0
python-dotenv==1.1.1
requests==2.32.5
pymongo==4.1.2